adds cost memory chart
